fix(ci): make jscpd exit non-zero on any clone
`website/.jscpd.json` sets no `exitCode`, and jscpd's own default is
`exitCode: 0`. Its CLI does `if (clones.length > 0) exitCallback(options.exitCode)`,
so the gate detects duplication, prints it, and exits 0 — it has never been able
to fail. Verified on main: `npx jscpd .` printed `Found 1 clones.` and
`Clone found (javascript):` with exit 0.

`threshold: 0` stays. It is not dead config: it is the separate path that still
fires on bulk duplication if `exitCode` is ever dropped again.

Landing the gate needs the one existing clone gone, so it goes with it. Both
topbar capture harnesses carried `diffPngs` — 19 lines, 367 tokens, differing
only by a two-line comment — so it moves to `scripts/lib/diff-pngs.mjs`
alongside the 32 helpers already there, and `TOL` becomes a parameter rather
than a captured constant: the tolerance is a property of the surface under
assertion, not of the diff, and both callers already declared their own.

Not taken: adding the pair to `.jscpd.json`'s `ignore`. That is the silencing
option, and the duplication is real.

Verified both directions after the change — clean tree exits 0 with
`Found 0 clones.`, and re-inlining the helper takes it to exit 1 with
`Found 1 clones.`
